+++ CHANGELOG.md
@@ -1,7 +1,14 @@
-## 1.27.1 (Unreleased)
+## 1.28.0 (Unreleased)
+
+FEATURES:
+
+* **New Data Source**: `tencentcloud_cbs_snapshot_policys`
+* **New Resource**: `tencentcloud_cbs_snapshot_policy_attachment`
+
 ## 1.27.0 (December 17, 2019)
 
 FEATURES:
+
 * **New Data Source**: `tencentcloud_tcaplus_applications`
 * **New Data Source**: `tencentcloud_tcaplus_zones`
 * **New Data Source**: `tencentcloud_tcaplus_tables`